production/buildbooth/buildbooth/requirements/base.txt
# Bleeding edge Django
#https://github.com/django/django/archive/1.6b4.tar.gz

# Configuration
django-configurations==0.5.1
django-secure==1.0
django-cache-url==0.6.0
dj-database-url==0.2.2

# Forms
django-braces==1.2.2
django-crispy-forms==1.4.0
django-floppyforms==1.1

# Models
django-model-utils==1.5.0

# images
Pillow==2.1.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.13.0

# For the persistance stores
psycopg2==2.5

# Unicode slugification
unicode-slugify==0.1.1
django-autoslug==1.7.1  

# There are edge cases for South with Django 1.5+ that haven't been addressed yet.
# South==0.8.1
https://bitbucket.org/andrewgodwin/south/get/59f6bae8b1a501ca14a5f23f8b11c44c42f33147.zip


# Useful things
# django-avatar that work with Django 1.5+
git+git://github.com/jezdez/django-avatar@6393d25166a6c2d2df0bd28e19f161fac2bb1166


# Your custom requirements go here
